+++ test/src/services/$Request.spec.js
@@ -6,7 +6,8 @@ import simple, { mock, spy } from   'simple-mock';
 import { Form } from                'multiparty';
 
 // Angie Modules
-const $Routes =                     require(`../../../${TEST_ENV}/factories/$RouteProvider`),
+const TEST_ENV =                    global.TEST_ENV || 'src',
+    $Routes =                       require(`../../../${TEST_ENV}/factories/$RouteProvider`),
     $Responses =                    require(`../../../${TEST_ENV}/services/$Response`),
     $Request =                      require(`../../../${TEST_ENV}/services/$Request`),
     $Util =                         require(`../../../${TEST_ENV}/util/Util`).default;
@@ -229,9 +230,10 @@ describe('$Request', function() {
                     }
                 }));
                 request = new $Request(req);
+                request.path = '/test';
                 mock(
                     $Responses,
-                    'ControllerTemplateResponse',
+                    'ControllerTemplatePathResponse',
                     function() {
                         throw new Error();
                     }