+++ test/src/util/Util.spec.js
@@ -3,7 +3,8 @@ import { expect } from              'chai';
 import simple, { mock } from        'simple-mock';
 
 // Angie Modules
-const Util =                        require(`../../../${TEST_ENV}/util/Util`),
+const TEST_ENV =                    global.TEST_ENV || 'src',
+    Util =                          require(`../../../${TEST_ENV}/util/Util`),
     $Util =                         Util.default,
     $StringUtil =                   Util.$StringUtil;